Move interfaces in separate files
