+++ src/sockTrader/core/bot/sockTrader.ts
@@ -1,12 +1,13 @@
 import uniqBy from "lodash.uniqby";
 import uniqWith from "lodash.uniqwith";
-import {ICandle, ICandleInterval} from "../candles/candleManager";
-import {IExchange} from "../exchanges/exchangeInterface";
 import {IOrderbook} from "../orderbook";
 import {IReporter} from "../reporters/reporterInterface";
 import BaseStrategy, {IAdjustSignal, ISignal, IStrategyClass} from "../strategy/baseStrategy";
 import {IOrder} from "../types/order";
 import {Pair} from "../types/pair";
+import {IExchange} from "../types/IExchange";
+import {ICandle} from "../types/ICandle";
+import {ICandleInterval} from "../types/ICandleInterval";
 
 export interface IStrategyConfig {
     interval: ICandleInterval;
