+++ src/sockTrader/core/exchanges/hitBTC.ts
@@ -1,13 +1,16 @@
 import crypto from "crypto";
 import nanoid from "nanoid";
-import {ICandle, ICandleInterval} from "../candles/candleManager";
 import {Data} from "../connection/webSocket";
 import logger from "../logger";
 import Orderbook from "../orderbook";
+import {IOrderbookData} from "../types/IOrderbookData";
+import {IResponseAdapter} from "../types/IResponseAdapter";
 import {IOrder, OrderSide} from "../types/order";
 import {Pair} from "../types/pair";
-import BaseExchange, {IOrderbookData, IResponseAdapter} from "./baseExchange";
+import BaseExchange from "./baseExchange";
 import HitBTCAdapter from "./hitBTCAdapter";
+import {ICandle} from "../types/ICandle";
+import {ICandleInterval} from "../types/ICandleInterval";
 
 export const CandleInterval: Record<string, ICandleInterval> = {
     ONE_MINUTE: {code: "M1", cron: "00 */1 * * * *"},
