+++ src/sockTrader/core/types/ICandle.ts
@@ -1,0 +1,10 @@
+import {Moment} from "moment";
+
+export interface ICandle {
+    close: number;
+    high: number;
+    low: number;
+    open: number;
+    timestamp: Moment;
+    volume: number;
+}
