+++ src/sockTrader/core/types/ICandleInterval.ts
@@ -1,0 +1,4 @@
+export interface ICandleInterval {
+    code: string;
+    cron: string;
+}
