+++ src/sockTrader/core/types/ICurrencyMap.ts
@@ -1,0 +1,5 @@
+import {ITradeablePair} from "./ITradeablePair";
+
+export interface ICurrencyMap {
+    [key: string]: ITradeablePair;
+}
