+++ src/sockTrader/core/types/IOrderbookData.ts
@@ -1,0 +1,9 @@
+import {IOrderbookEntry} from "../orderbook";
+import {Pair} from "./pair";
+
+export interface IOrderbookData {
+    ask: IOrderbookEntry[];
+    bid: IOrderbookEntry[];
+    pair: Pair;
+    sequence: number;
+}
