+++ src/sockTrader/core/types/IResponseAdapter.ts
@@ -1,0 +1,5 @@
+import {Data} from "../connection/webSocket";
+
+export interface IResponseAdapter {
+    onReceive(msg: Data): void;
+}
