+++ src/sockTrader/core/types/ITradeablePair.ts
@@ -1,0 +1,7 @@
+import {Pair} from "./pair";
+
+export interface ITradeablePair {
+    id: Pair;
+    quantityIncrement: number;
+    tickSize: number;
+}
