+++ src/sockTrader/core/types/exchanges/IHitBTCAuthenticateResponse.ts
@@ -1,0 +1,5 @@
+export interface IHitBTCAuthenticateResponse {
+    id: string;
+    jsonrpc: string;
+    result: boolean;
+}
