+++ src/sockTrader/core/types/exchanges/IHitBTCGetSymbolsResponse.ts
@@ -1,0 +1,14 @@
+export interface IHitBTCGetSymbolsResponse {
+    id: string;
+    jsonrpc: string;
+    result: Array<{
+        baseCurrency: string;
+        feeCurrency: string;
+        id: string;
+        provideLiquidityRate: string;
+        quantityIncrement: string;
+        quoteCurrency: string;
+        takeLiquidityRate: string;
+        tickSize: string;
+    }>;
+}
