+++ src/sockTrader/core/types/exchanges/IHitBTCOrderbookResponse.ts
@@ -1,0 +1,12 @@
+import {IOrderbookEntry} from "../../orderbook";
+
+export interface IHitBTCOrderbookResponse {
+    jsonrpc: string;
+    method: string;
+    params: {
+        ask: IOrderbookEntry[],
+        bid: IOrderbookEntry[],
+        sequence: number,
+        symbol: string,
+    };
+}
