+++ src/sockTrader/data/candleNormalizer.ts
@@ -1,9 +1,9 @@
 import {IDataFrame} from "data-forge";
 import {IAsyncFileReader, readFile} from "data-forge-fs";
 import moment from "moment";
-import {ICandle} from "../core/candles/candleManager";
 import {Pair} from "../core/types/pair";
 import {getDecimals} from "./utils";
+import {ICandle} from "../core/types/ICandle";
 
 export interface ICandleNormalizerConfig {
     decimalSeparator?: string;
