+++ src/strategies/simpleMovingAverage.ts
@@ -1,12 +1,11 @@
 import {sma as SMA} from "technicalindicators";
 import CandleCollection from "../sockTrader/core/candles/candleCollection";
-import {IExchange} from "../sockTrader/core/exchanges/exchangeInterface";
-import logger from "../sockTrader/core/logger";
 import {IOrderbook} from "../sockTrader/core/orderbook";
 import BaseStrategy from "../sockTrader/core/strategy/baseStrategy";
 import {crossDown, crossUp} from "../sockTrader/core/strategy/utils";
 import {IOrder, OrderStatus} from "../sockTrader/core/types/order";
 import {Pair} from "../sockTrader/core/types/pair";
+import {IExchange} from "../sockTrader/core/types/IExchange";
 
 /**
  * Strategy using simple moving average
